+++ src/vaults/UltraVault.sol
@@ -327,7 +327,7 @@ contract UltraVault is BaseControlledAsyncRedeem, UUPSUpgradeable, IUltraVaultEv
 
         _getUltraVaultStorage().proposedFundsHolder = AddressUpdateProposal({
             addr: newFundsHolder,
-            timestamp: block.timestamp
+            timestamp: uint96(block.timestamp)
         });
 
         emit FundsHolderProposed(newFundsHolder);
@@ -364,7 +364,7 @@ contract UltraVault is BaseControlledAsyncRedeem, UUPSUpgradeable, IUltraVaultEv
 
         _getUltraVaultStorage().proposedOracle = AddressUpdateProposal({
             addr: newOracle,
-            timestamp: block.timestamp
+            timestamp: uint96(block.timestamp)
         });
 
         emit OracleProposed(newOracle);
